dom/nodes/Node-appendChild-script-and-custom-from-fragment.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and custom element from a DocumentFragment</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-node-appendChild">
<link rel=help href="https://dom.spec.whatwg.org/#concept-node-insert">
<link rel=help href="https://github.com/whatwg/dom/issues/575">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
var script = document.createElement("script");
var scriptRan = false;
script.textContent = `
    customElements.define("custom-element", CustomElement);
    assert_true(customConstructed, "custom element was not constructed");
    scriptRan = true;
`;

var custom = document.createElement("custom-element");
var customConstructed = false;
class CustomElement extends HTMLElement {
    constructor() {
        super();
        customConstructed = true;
    }
}

var df = document.createDocumentFragment();
df.appendChild(script);
df.appendChild(custom);

assert_false(scriptRan, "script ran");
document.head.appendChild(df);
assert_true(scriptRan, "script has not run");
done();
</script>
